feat: landing 페이지 폴더 삭제 (common 그룹의 page로 내용이동)
