+++ src/app/(common)/_components/Banner.module.scss
