+++ src/app/(common)/_components/Banner.tsx
