+++ src/app/(common)/_components/landingContents.ts
