+++ src/app/(common)/page.tsx
@@ -1,5 +1,17 @@
-import LandingPage from "./_landing/page";
+import Banner from "./_components/Banner";
+import Feature from "./_components/Feature";
+import {
+  heroBannerData,
+  footerBannerData,
+  featureList,
+} from "./_components/landingContents";
 
-export default function Page() {
-  return <LandingPage />;
+export default function LandingPage() {
+  return (
+    <>
+      <Banner {...heroBannerData} />
+      <Feature list={featureList} />
+      <Banner {...footerBannerData} />
+    </>
+  );
 }
